+++ src/components/MyPage/Menus/Content.tsx
@@ -5,7 +5,7 @@ import SearchIcon from "@mui/icons-material/Search"
 import { useContext, useRef, useState } from "react"
 import { ThemeContext } from "../../../pages/_app"
 import { ThemeProps } from "../../../types/Theme"
-import { ContentData } from "./ContentData"
+import ContentData from "../../Common/ContentData"
 import { useData } from "../../../hooks/useData"
 import qs from "qs"
 import { Theme } from "../../../styles/theme"
